[APPLE-920] INTERIM: converting to token-based parser
THIS DOES NOT RUN CORRECTLY.  It compiles,
and runs, but it always delivers the wrong result.

In the schedule-expression parser, removed lots
of redunancy, by moving some logic "upwards" in
the parser:  any sub-expression can now contain
any of the legal tokens, and only at the last minute
do I decide which expression is (say) "months" and
which is (say) "weekdays.")

However, to do this, thee previous character-based
parser and new token-based parser are both parsing
THE SAME TOKENS, and the char-based parser
isn't "consuming" the token-based tokens.

To fix this, I need to either
(a) make the char-parser consume the
     token-parser's tokens, or
(b) convert everything to the token-based parser.

I'm not yet sure which I'm going to do, so I'm
committing this branch.