[APPLE-920] ScheduleExpression now handles weird start dates.
Two discussions here:  what I did, and how I did it.

Short version:
- to handle a special edge case with start dates, I
- refactored all the enumeration logic, so that
- there's only a single enumeration method, not two.
- I also removed dependencies on sub-enumerators, and
- added NSDate+Helper methods.

I tested this with many combinations of expressions
and start/end dates in -testRealisticTests.

Long version:

---
What I Did
---
The ScheduleExpression enumerator now handles
a specific edge case for start dates:  when the day
value for the start date is later in the month than the
last day in the list of legal days in the month.

For example, if the legal days contain only "the first
Friday of every month," and the start date is "March 15,"
the first legal day is the first Friday in April, not
March.

Earlier versions of this code thought this situation would
never arise, and returned nil.

---
How I Did It
---
Removed the redundancies and somewhat different
algorithms in -init and -nextObject.  Specifically:

- The -init method now does almost nothing:  sets up
some variables and exits.  It now looks like a "normal"
init method.

- The -nextObject method now returns either -firstDate
or -nextDate, according to whether or not this is the first
time it was called.

- All the brains of the iteration are now in -nextDate.
Previously, -init and -nextObject had different ways of
thinking about the iteration and the date components --
two different algorithms.  All that logic is now streamlined
and lives in -nextDate.

- The -firstDate method handles the "edge case" of
setting things up for the first time.  To handle the
edge case about start dates which was the point of
this commit, it creates a start date from the enumeration
rules, and then simply calls -nextDate until it hits or
passes the startDate specified by whoever's using
the enumerator.

- This will also enable us to create a -reset method
for the enumerator if we like.  I didn't do that, because
we don't currently need it, but it will be literally trivial:
set -hasStartedYet to YES.

During that refactoring, I found I could REMOVE ALL
DEPENDENCIES ON OTHER ENUMERATORS.
This ScheduleEnumerator now asks each Selector for
its next or initial value.  It doesn't use
TimeSelectorEnumerators wrapping each selector,
as did the previous versions.  The resulting code seems
more streamlined and focused.

I also added a set of NSDate helper methods in
NSDate(Helper).