[APPLE-920] Rearranged code that got separated during merge.
When I merged Dev into this branch, I had
a merge conflict.  In fixing it, I separated some
code that should have been logically grouped
together. This just cuts-and-pastes that code
into the right place (clustered with similar
methods, and arranged in the same order
as in the header file).