[APPLE-1243] Got original scripts from Dhanush and Ed.
Created a new directory in AppCore:  LocalLoggingServer.

Put an "originals" directory in there, with the original
logging-server experiments we did, as a team, a few
months ago:
- Dhanush's Ruby-based server
- Ed's copy of Apple's Python-based server from
   an older copy of ResearchKit

I then copied Dhanush's "original" server file to the root
of this new directory, calling it:  localLoggingServer.rb.

Making this run on my machine required these commands:
- sudo gem install sinatra
- sudo gem install colorize
- sudo gem install pry

Those last 2 lines are optional, as long as you comment
out the matching "require" lines at the top of the .rb file.