[APPLE-1272] Fixed crash on forgot-password page.
The problem was an "assertion failure" in
-[UIKeyboardTaskQueue waitUntilAllTasksAreFinished],
which is Apple's method, not ours.

I fixed 2 things.  After discussion with Ed and Farhan,
we're not sure why they *actually* worked (but they do
indeed work perfectly):

- Moved a Flurry log statement from "viewDidDisappear"
to "viewDidAppear" (where it belonged).  Doing this made
the crash go away.  This is odd.  It suggests that the crash
was because of a timing issue, which suggests we haven't
finished finding all the bugs related to that issue.

- Moved a "dismissViewController" from a background
thread to the main thread.  On the surface, this makes
perfect sense...  except if that's the case, why does
moving the above logging statement also solve the
problem?

Also:

- reorganized the asynchronous stuff in -sendPassword
so that pieces seem more closely associated with
appropriate other pieces.

- added a localization comment.