[APPLE-1243] Removed original code from Dhanush and Ed.
Removing this code, which is actually unused,
from the branch.  This is the code on which
all my work in this branch was based:

- Dhanush's original Ruby server

- Ed's copy of Apple's original test server,
which he got from Apple in an early version
of the Research Kit.  (Actually, I never used
this code, but it's still the real original code.)

If we need to resurrect this code, we can
always come back to a commit before now
and reinstate it.