[APPLE-1243] Made scheme for sending data to server.
Made an Xcode "scheme" that compiles and uses the
code which talks to our DataValidationServer.  (This is
the stuff which sends our unencrypted .zip files to a
local server, which I wrote, also on this branch.)

The stuff in this commit contains just the part for
AppCore. It's safe to use this without the parent app
or the Scheme; the code in question will simply be
compiled out.

Short version of what I did (details below):

- Created a new Configuration that defines a macro
our code can access
- Created a Scheme which launches that Configuration
- Added #ifs to the code, using that macro

Details:

- In both the main app (Diabetes) and AppCore, clicked
on the project file, and then the project entry, and created
a new Configuration called
"DebugAndSendToDataVerificationServer", as a duplicate
of the Debug configuration.

- In AppCore, clicked on the project file, then then project
entry, then scrolled down to "Preprocessing," and added
the preprocessor macro
USE_DATA_VERIFICATION_CLIENT=1 to the
preprocessor entry for our new Configuration.

- In Xcode, created a new Scheme, duplicating the
existing Diabetes scheme, called it
DiabetesUsingDataVerificationServer, and made it
"Shared."

- Finally, down in AppCore, wrote #if statements using the
USE_DATA_VERIFICATION_CLIENT flag to enable or
suppress any code that exposes this "send unencrypted
data to a server" feature. This way, the code is both
invisible and inaccessible to people trying to hack into
the app.