[APPLE-1243] Added special file for personal IP addresses.
Made a special easy-to-find file where we can
add our own IP addresses, where we can find
the DataValidationServer.

As with the other files in this package, this file
will get compiled out if
USE_DATA_VERIFICATION_CLIENT is not set --
i.e., if we're not using the Scheme called
"...UsingDataValidationServer."

Also added an error message which appears if
we can't find the DataValidationServer.  It
says, "please make sure you're on the office
local network."