[APPLE-1243] Numbering the printed files.
...as a way of helping make the reader's life
easier.  These files are still in a meaningless
sequence -- that's determined by the code
generating the files -- but this makes it
(hopefully at least somewhat) easier
to move your eyes from the list of all files to
the contents of a single file without getting
lost in the endless stream of text.