[APPLE-1571] Added app and phone info to Sage data.
Added three items to the Sage data, in the
"info.json" file:

- app name (like "mPower")
- app version (like "version a22, build 1062")
- phone info (like "iPhone 6 4G" or "Simulator")

For all of them, I used utilities or calls that other
people had written elsewhere in AppCore.

I also have code in there which generates
a human-readable time stamp, like

    2015-02-05 13:36:02 PST (Thursday,
    February 5th, 1:02 PM, Pacific Standard
    Time)

but I commented it out, 'cuz it may be
unnecessary.