[APPLE-1557] Fixed newlines in JSON data, I think.
Formerly:  we were stringifying everything except strings and numbers.

This commit allows arrays and dictionaries through -- as long as they contain only strings, numbers, and nulls inside them, or other arrays/dictionaries of such primitives.  This is because we're using NSJSONSerialization to convert to JSON, and those are the only things it allows.

This commit also contains a bunch of #warnings, as questions/notes to myself about how this works.  I'll remove those before I commit this to dev.

Lastly, this commit contains one renamed method, which (to me) more accurately reflects what it does.  In doing that, I found one call to that method which seems to be using it in a way it wasn't intended.