[APPLE-1557] Systematically gathering property names.
Previously, we had three lines of code which collected the names of properties we wanted to store in JSON, for a given object.

I replaced those three lines with a loop that crawls up that object's superclasses, to make sure we don't miss any superclasses.  As implied by the previous code, I stop the crawl when we reach the class RKSTResult, which seems to be the top superclass we care about recording.