[APPLE-1557] Removed side effects from a method.
Previously, we had a method with lots of "side effects":  we passed it a MutableDictionary and modified that dictionary.  But then we made a COPY of all the stuff in the dictionary, further modified the copy, returned the copy, saved the copy to disk, and sent the copy to Sage.

This method is used in exactly two places.  One place is central to the DataArchiver class; the other is a specialty method used by Parkinson's.  It's used nowhere else in the code base.

After all that analysis, it seemed to me that the method's primary function was to generate serializable data, not modify the source data.  So I modified the method as follows:

- renamed it to reflect its purpose-without-side-effects.  It was -processDictionary.  It's now -generateSerializableDataFromSourceDictionary.

- changed its parameter to a non-mutable dictionary, to remove those side effects.

Note:  I haven't yet tested the Parkinson's feature which uses that specialty method.  I'll do that next.