[APPLE-1557] Centralized: converting RK types to strings.
Previously, we had a hard-coded array which contained NSString versions of the RKQuestionType enum.  However, the code didn't explicitly refer to that enum.

This commit contains a Category on RKSTAnswerFormat which contains a method for converting RKQuestionType to a string.  It's still hard-coded.  However, I hope that by putting it in a central place, named after the content it's actually translating, it'll help us understand and maintain it.

I put it in RKSTAnswerFormat because that's where RKQuestionType is defined.  I think that makes it hard to find, though.  If you have a better name for the category, or a better place for that method, please change it.

To address that "gee, this is hard to find" problem, I made a standard-looking, Objective-C style NSStringFromRKQuestionType() function, which calls that category method.