[APPLE-1557] Converting stringified ints and bools to real values.
A core feature of ResearchKit seems to be:  when Task questions are multiple-choice questions (either multi-select or single-select), the answers are always strings, even if they came from some original JSON file as integers.  Same goes for Booleans (represented as "no", "yes," "true," and "false").

So this commit:
- looks for arrays we're about to serialize
- crawls through them looking for integers and booleans
- converts them to real integers and booleans
- wraps them in NSNumbers
- and serializes those numbers.

So, now, our serialized arrays look like [1, 2, 3], instead of ["1", "2", "3"].

That was a major point of this whole branch.  Before this branch, the above arrays looked like "(\n   1,\n   2,\n   3,\n)" (including the literal back-slashes).  Previous commits in this branch converted that string to an actual array.  This commit converts the stuff IN the array to its real values.

In addition, this commit pulled a lot of "safe-ification" methods into a separate section at the bottom of the file, to make the core serialization method -generateSerializableDataFromSourceDictionary easier to read (I hope).