[APPLE-1557] Removing a fix for an unrelated bug.
This commit undoes commit 5373e5c27fd9242e8e562d7e3f1bd0fdaa7e1177.  I added that fix to this branch because I couldn't progress without it.  But now (a) we have another JIRA ticket for that (APPLE-1780), and (b) I know how to work around it.  So I'll unsully this branch, and then check that in on its own branch.