[APPLE-1780] Preventing crash when receive "unreasonable" height.
Before this fix:  if the user supplied an "unreasonable" height in the Health system on the phone, the app would crash.

This fix prevents the crash:  if the height is outside the known list of heights, we'll choose a "default" height (5 feet tall).

Please note:  this fix is only part of the fix to a bigger bug.  The bigger bug is that our apps will *increase* the user's height each time we reinstall the app.  After enough reinstalls, the height becomes "unreasonable," and this secondary bug kicks in, causing the app to crash.  This commit fixes the bug which causes the crash, but not the bigger bug which increases the user's height.