[APPLE-1697] Making AppCore support test suite in main app.
Exposing various features down here in AppCore so that my CoreData test cases for the Medication Tracker can be run from the higher-level app — in this case, Parkinson's.

We have to run the tests in the main app because, otherwise, we don't have access to the CoreData infrastructure features (the APCDataSubstrate) which get instantiated up there.

I'll commit the matching Parkinson's changes in a moment, in a branch with the same name.  This AppCore commit can be used without Parkinson's, but the Parkinson's commit will require this version of AppCore.