[APPLE-1697] Added dateStarted to the medication schedule.
Records the date the patient started and stopped using a certain medication, by attaching those dates to a Schedule.  Also records whether the patient stopped on a doctor's orders (currently the only way to do it, so that may be gratuitous).

To make Xcode regenerate the class file, I had to delete it first, which is why that file seems to have been removed from and re-added to the project file.

Also added some utility methods on NSDate.

This commit is independent of the main app.  I'm about to commit a version of Parkinson's  which depends on this version of AppCore, though, for these new features.