[APPLE-1697] Made a date optional. Found bugs in our DataSubstrate.
Primarily, made a Date field optional.

Secondarily, but significant:  Farhan and I think we found that our CoreData code isn't reporting errors like this.  Among other things, it's absorbing data-validation errors...  such as

- making a certain attribute "required" in CoreData, and then
- forgetting to set that value when you instantiate the object in question.

The problem is that the app threw an exception which was hard to understand; you have to unpack it a little to realize it's saying:  "you didn't handle a CoreData error!"

I'm about to commit some Parkinson's code which depends on this.