[APPLE-1697] Reading predefined colors from disk.
...as a way of getting data into the database in the first place, for the user to choose from.

This also includes a first pass at the logic for making sure I don't generate duplicate colors of the same name, and (thus) allow us to update colors by changing the values in the .plist file on disk.

Renamed some of the .plist files to work with this new scheme.  This means they won't work with the old code which read them.  I haven't yet deleted that code, though.

About to check in a version of Parkinson's that depends on this commit.