[APPLE-1697] Superclassed some of my model objects.
Preparing to generalize the ability to load these specific types of objects from disk.  I gave them a common superclass, in the data model file, and regenerated the class files for each of those classes.  Then I re-ran a bunch of the test cases to make sure everything still works.  It does.