[APPLE-1697] Renamed "MedicationSchedule" to "Prescription."
This was tricky.  I had to do a lot of it manually.  Argh.  But repeated tests seem to show this works fine.  However:  if you have an existing CoreData database based on this, you'll have to delete it.  Not yet sure why, but since we haven't released this, it doesn't matter.