[APPLE-1697] Recording and fetching DailyDosageRecords.
Changed the CoreData entity "ActualDosageTaken" to a "DailyDosageRecord."  I.e., we're not storing and retrieving individual doses of medication that a user takes; we're recording, merely, how many doses the user has taken on a given date.  Tracking anything more than that is unnecessary engineering, as well as being misleading -- we don't and won't know when the user actually takes her pills.

I'm about to submit a matching test case in Parkinson's.