[APPLE-1697] Making my operation-queue priority explicit.
I wrote a category on NSOperationQueue which blindly accepted the system's default priority.  Now I'm setting it explicitly — to the same thing, for now, until we get a chance to play with it.  Wrote a generalized method to set the priority however we want, but I'm not exposing it yet, because I don't want people accidentally setting the priority to "high" and competing with the main thread.