[APPLE-1162] Forcing Parkinson's JSON files to look like "JSON."
This is an experiment.  I hard-coded the Archiver to look for "known JSON filename prefixes" -- any files that we see ResearchKit generating which we just happen to know to be JSON -- and then I'll remove "."s from their filenames, and attach the extension ".json".

Potential problems:  Not scalable or future-compatible.  If whoever came up with these filenames decides to change them, this breaks.  If another app uses one of these filenames for something which is *not* JSON, this *might* break.  Let's think about this.

Still, this does indeed work.  Files which used to look like

    accel_walking.outbound-20150225191827

are now coming out as

    accel_walking_outbound-20150225191827.json