Declared and used a new APCProfileViewControllerDelegate method that will allow the profile view's delegates to know when it is editing -- a notion that is important in order for the profile view's rows to have the proper highlighting behavior.  This also requires that apps providing this delegate to use these methods.
Parkinson, in particular, has such a delegate.  Before I decided to add these delegate methods, I noticed the delegator was a property in the public header for this delegate.  In my search, however, I could not find any reference to using this delegator.  That is the reason why I added these new delegate methods.  This way, the delegate has the option of receiving messages about the profile's editing without needing to declare the profile view (the delegator) in its public header as a property to be set.