[APPLE-2122] Added data-model migration.
Consider this an experiment.

The "bug" was to address what happens if we can't create or open the user-data database.  Before, we were simply deleting the data and recreating it.

The catch is:  we still don't have a solution.  If it's not acceptable to do blow away the user's data, but we STILL, ACTUALLY, cannot create or open the database, what do we do?  Tell the user to give up and go away?

So this commit still has that problem.  It's even "worse":  there are two NSAsserts in the code, to make sure we think about this in the next couple of days.

However, I've added a bunch:

- I report separate errors for 3 situations:
  - open existing database, and fail
  - create new database, and fail
  - create new database, and fail, after trying to open existing

- Each of those errors has human-readable text in it, and contains the CoreData error as an "underlying error."

- I improved the APCLog printout to make "underlying errors" print more readably.

- THE BIG DEAL:  I turned on data migration.

This was, apparently, the reason we implemented "delete the user database and try again."  I revved the database model, using Editor > Add Model Version. I made a slightly non-trivial modification to the database -- adding an Attribute to a column in part of the model I'm responsible for.  I told Xcode to use my new model as the current version.  Then I ran the app.  It correctly upgraded the existing database and plowed onward.  I'm able to use the MedTracker to store and retrieve data.  I did this twice.  It works.

But that still doesn't address the end-user problem of "can't open the user's previous database," for some OTHER reason.

It also adds complexity.  We now have to remember to always create a "new model version," make that the "current version," double-click the appropriate file, and make our changes THERE.