[APPLE-2122] Removed "auto-delete existing data."
Deleted the code which deletes the user's data if we can't open the existing database.

Now, if we can't open the user's existing database, even by converting it, I'll log the error and... um... call NSAssert.  Which does nothing in production.  In a later commit, I'll figure out how to get something onto the UI.