[APPLE-2231] Logging and ignoring duplicate passive data collector.
This is a hard-to-reproduce problem:  sometimes, we get a duplicate entry for a Passive Data Collector.

The "fix":  use the existing data collector, not the new one.

But the real fix would be:  clean up the collectors when the user does something appropriate, like tapping "Leave Study."

I looked for a way to do that, but I didn't see anywhere that ANY collectors are being cleaned up.  (I'm also new to the Collectors, so I could be missing it.)