[APPLE-2479] Incorrect Logging of Medications Taken
Ron here, at Henry's machine:  Henry found, and I fixed, a couple of bugs in the CoreData layer.  Ahem.  (sound of whistling, glance upward, look off in the distance)  When we retrieved, created, and deleted DailyDosageRecords from within a Prescription, I was creating/destroying ALL records for ALL prescriptions, every time, instead of the records for the prescription we thought we were operating on.