[APPLE-2526] Logging files being archived and uploaded to the Console.
This is NOT going to Sage; this is only going to the Console, for our debugging information.

To see this stuff in the console, search for this regular expression:

    apc_(view|archive|upload)

This will walk you through each view that appears, and then the archives (.zip files) and uploads that happen while that view is on-screen.