[APPLE-2524] FIRST PASS at "please check your email" alert.  DO NOT USE.
I say "first pass" because there are many edge cases here, any one of which might reveal itself to be the "normal" situation -- and any one of which could mean the user might NEVER GET THROUGH THIS SCREEN.

That being said:  this is a first pass at displaying a dialog prompting the user to check her email and click the email-verification link.

This is dangerous and edge-case-y because it introduces a modal dialog into something with a bunch of asynchronous states, any one of which might fail if the dialog is showing.  Or, as we found, even if it's NOT showing.

So I'm committing this because it was a lot of work, but I don't think we should ship it.

As Farhan and I discussed, a better solution to this might be:  redesign this screen, taking into account all of the states it goes through, all the asynchronous server-communications callbacks and local callbacks it triggers, and all the threads those things go onto.  Then make a centralized method responsible for managing changes between states:  a method which knows, for certain, which alerts are on the screen or not, and which asynchronous processes are in progress and which aren't.  Then make everything on the screen call that method to change state, instead of changing state in some obscure "if" statement somewhere.