Medication Tracker Fixed the Random Symptoms
Ron here, at Henry's machine:  Henry found a bug in the data-storage layer -- I wasn't actually deleting any Dosage Records when I claimed to.  Ahem.  This fixes that.  As a result, in the UI, when you update the number of pills you've taken on a given day, the results now look correct.