[APPLE-2693] First version of a working uploader!  Woo-hoo!
Contains a working uploader, and modifications to MedTrackerPrescription which uses this to upload stuff to Sage.  It's working!

It also has some issues:

(1) The info.json file (the upload manifest) is missing the "item" entry, which says what the whole upload is for.

(2) I rewrite filenames to be "safe," but didn't make the JSON copy of the filename match.

(3) I haven't tested the edge cases or "unhappy paths" at all.