[APPLE-2693] Working uploader. Haven't checked all error cases.
Contains a rewriting and streamlining of the previous commits from this branch, based on Ed's feedback.  (This is actually a new branch with a bunch of stuff from earlier experiments left out.)

This new ArchiverAndUploader class actually works -- data gets sent to Sage, and copied to my data-verification server.  I have not yet checked all the error cases in this class, although I've set them all up in a parallel way, so they'll hopefully be easy to debug and enhance.

Also:
- created a large set of public error constants -- error codes and matching strings -- in APCConstants.h
- thanks to those errors, I cleaned up (a little bit) my error-handling in DataSubstrate.  I haven't tested that; I just swapped in the new error methods and constants.
- created several utility methods for generating detailed errors easily