[APPLE-2712] Made encryption kinda optional.
This commit lets you turn encryption on or off by choosing to include one of two possible .m files in the project.  There are 2 files:  noEncryption and openSSL.

-  If you include the noEncryption file in the APCAppCore target (click its checkbox in Xcode's File Inspector, and un-check the checkbox next to the openSSL file), encryption is effectively turned off.

-  If, instead, you include the openSSL file in the APCAppCore target, encryption will be turned on.

In the next commit, I'll make sure we can also physically remove the openSSL library from the project if we choose the noEncryption option.