[APPLE-2814] Coercing "7" for day of week into "0".
I used to do this.  Must'a lost it somewhere along the evolution of this project.

Added this "coerce" method in the parser, where I think we can most reasonably catch it.  However, it's ambiguous.  It should more properly be handled in a subclass called (something like) APCWeekdaySelector.  We don't currently have *any* date-field-specific subclasses like that, as part of the original and ongoing design.