[APPLE-2796] The uploader now takes ownership of uploaded files.
The uploader now assumes physical control of the files you ask it to upload.  When you ask it to upload a file (or several files), it moves them to a temp directory *while on the thread from which you called it*, so that your next line of code can be to do something with the path(s) where those files used to be.  It returns YES or NO depending on whether it was able to move the files, and provides an error object if you ask for it.  If the "move" worked, it will then start to upload the files themselves on a separate thread.