[APPLE-2846] Fixed the crash when sharing on Twitter.
The crash was due to an older Twitter client that was installed. Aparaently iOS uses the offical Twitter client if it is installed; otherwise it uses it's internal interface for posting to Twitter.

In order to post to Twitter the client must have been updated to include app extension to allow posting from other apps. WIthout it the composer API will return nil and crash app. Therefore we are now gurading against that possibility.