[APPLE-2796] Fixed uploader to add 'item' to info.json.
Whoops.  Also adds the "taskRun" field.  Both fields are fields added to the info.json file in Dhanush's original archiver.  Didn't realize they were significant.

This came up in discussion with Sage, when they very politely asked us what value we were planning to put in the "item" field.

This commit changes the +upload methods to require a taskIdentifier and a taskRunUUID.  It'll accept nil or garbage for both of them, but it'll emit an error if the taskIdentifier field is nil.

Contains a change to the MedicationTracker's Prescription class, because that class had a reference to the older version of my +upload methods.