[APPLE-2873] - Demographic data (collected during onboarding) is not uploaded to Bridge
Created a new  APCDemographicUploader  class to access user data and upload the data.

Added a call to  APCAllSetContentViewController  to upload the demographic data after initial sign up.

Added code to  APCProfileViewController  to upload demographic data when edits are made (somewhat blunt instrument right now)

Added code to  APCAppDelegate  to check if there has been a previous demographic upload, and to perform the upload if it has never been performed before.

Signed-off-by: Henry McGilton <henry.mcgilton@ymedialabs.com>

# Conflicts:
#	APCAppCore/APCAppCore/Startup/APCAppDelegate.m
#	APCAppCore/APCAppCore/UI/Onboarding/SignUp/APCAllSetContentViewController.m
#	APCAppCore/APCAppCore/UI/TabBarControllers/Profile/APCProfileViewController.m