[APPLE-2796] Removed data-validation on taskIdentifer and taskRunUuid.
Realized:  we don't know how to validate these, yet.  So we'll just upload them however the programmer passes them to us.  If they're invalid values, that'll spark a conversation between Sage and the programmer using this tool, which will solve the real problem.  We can add data-validation methods later if we need 'em.