[JIRA-2796] Uploader now adds "item" field to JSON output, identifying the purpose of the upload.
This commit changes the new +upload methods to require a "taskIdentifier" string, as requested by the server team.  Also requires a "taskRunUuid," because elsewhere in the app, those two field seem be tied together -- one says what we're uploading (like "monthly heart-health survey"), the other says which copy of that data is being uploaded.

Stuff I fixed as a side effect:
- enhanced matching documentation
- fixed the MedicationTracker's Prescription class to use the new +upload method correctly (or else the app wouldn't compile)